reduce code duplication
