+++ Code/ForceField/CMakeLists.txt
@@ -1,21 +1,24 @@
 
 rdkit_library(ForceField
-              ForceField.cpp 
+              ForceField.cpp AngleConstraint.cpp
               UFF/AngleBend.cpp UFF/BondStretch.cpp UFF/Nonbonded.cpp
               UFF/Inversion.cpp UFF/TorsionAngle.cpp
-              UFF/DistanceConstraint.cpp UFF/AngleConstraint.cpp
+              UFF/DistanceConstraint.cpp
               UFF/TorsionConstraint.cpp UFF/PositionConstraint.cpp
               UFF/Params.cpp 
               MMFF/AngleBend.cpp MMFF/BondStretch.cpp MMFF/StretchBend.cpp
               MMFF/OopBend.cpp MMFF/TorsionAngle.cpp MMFF/Nonbonded.cpp
-              MMFF/DistanceConstraint.cpp MMFF/AngleConstraint.cpp
+              MMFF/DistanceConstraint.cpp
               MMFF/TorsionConstraint.cpp MMFF/PositionConstraint.cpp
               MMFF/Params.cpp
               LINK_LIBRARIES Optimizer Trajectory)
 target_compile_definitions(ForceField PRIVATE RDKIT_FORCEFIELD_BUILD)
 
 rdkit_headers(Contrib.h
-              ForceField.h DEST ForceField)
+              ForceField.h
+              AngleConstraint.h
+              Params.h
+              DEST ForceField)
 
 rdkit_headers(UFF/AngleBend.h
               UFF/BondStretch.h
